# backend/src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: DayPoo API
  description: 💩 DayPoo - 실시간 화장실 찾기 및 건강 리포트 AI 서비스 API
  version: v1.0.0
servers:
  - url: http://localhost:8080
    description: Local Server
paths:
  /api/v1/auth/signup:
    post:
      tags:
        - Authentication
      summary: 회원가입
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignUpRequest"
      responses:
        "200":
          description: OK
  /api/v1/auth/login:
    post:
      tags:
        - Authentication
      summary: 로그인
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
...
  /api/v1/admin/toilets/sync:
    post:
      tags:
        - Admin
      summary: 공중화장실 데이터 동기화 (행안부 API 연동)
      description: "[ADMIN 전용] 행정안전부 API로부터 최신 화장실 데이터를 가져와 DB와 Redis를 동기화합니다."
      security:
        - BearerAuth: []
      parameters:
        - name: pageNo
          in: query
          schema:
            type: integer
            default: 1
        - name: numOfRows
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: 성공
          content:
            text/plain:
              schema:
                type: string
        "403":
          description: 권한 없음 (ADMIN 전용)
...
  /api/v1/admin/stats:
    get:
      tags:
        - Admin
      summary: 시스템 통계 조회
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStatsResponse"
  /api/v1/admin/inquiries/{inquiryId}/answer:
    post:
      tags:
        - Admin
      summary: 1:1 문의 답변 등록
      security:
        - BearerAuth: []
      parameters:
        - name: inquiryId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminInquiryAnswerRequest"
      responses:
        "200":
          description: OK
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SignUpRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        nickname:
          type: string
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
...
    AdminStatsResponse:
      type: object
      properties:
        totalUsers:
          type: integer
        todayNewUsers:
          type: integer
        totalRecords:
          type: integer
        todayRecords:
          type: integer
    AdminInquiryAnswerRequest:
      type: object
      properties:
        answer:
          type: string
...